projects/dl/imaging/medical-image-diffusion/pyproject.toml
[project]
name = "med-diffusion"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.12.0",
    "albumentations>=2.0.8",
    "diffusers>=0.36.0",
    "hydra-core>=1.3.2",
    "ipykernel>=7.1.0",
    "matplotlib>=3.10.8",
    "numpy>=2.4.0",
    "omegaconf>=2.3.0",
    "opencv-python>=4.11.0.86",
    "pandas>=2.3.3",
    "pillow>=12.1.0",
    "pyarrow>=22.0.0",
    "pytorch-lightning>=2.6.0",
    "pyyaml>=6.0.3",
    "rich>=14.2.0",
    "safetensors>=0.7.0",
    "scikit-image>=0.26.0",
    "scipy>=1.16.3",
    "torch>=2.9.1",
    "torchaudio>=2.9.1",
    "torchmetrics>=1.8.2",
    "torchvision>=0.24.1",
    "tqdm>=4.67.1",
    "transformers>=4.57.3",
    "wandb>=0.23.1",
]
